job-load/requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
durationpy==0.9
google-auth==2.37.0
idna==3.10
kubernetes==31.0.0
oauthlib==3.2.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
urllib3==2.2.3
websocket-client==1.8.0
